# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ayu-Trial Fabric - Core Intelligence & Governance API",
    "description": "Part B Backend & Governance Compiler API for Ayu-Trial Fabric. Powers ChangeSet impact resolution, deterministic rule evaluation, obligations lifecycle, evidence verification, and immutable audit tracking.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/v1/compiler/run": {
      "post": {
        "summary": "Execute the governance compiler against a ChangeSet",
        "description": "Evaluates rules, findings, obligations, and returns an immutable CompilationRun record.",
        "operationId": "runCompiler",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompilationRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Compilation run results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompilationRunResponse"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntityError"
          }
        }
      }
    },
    "/api/v1/compiler/findings": {
      "get": {
        "summary": "Retrieve findings for a ChangeSet or compilation run",
        "operationId": "getFindings",
        "parameters": [
          {
            "name": "changeSetId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of findings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compiler/obligations": {
      "get": {
        "summary": "Retrieve governance obligations",
        "operationId": "getObligations",
        "parameters": [
          {
            "name": "changeSetId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of obligations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compiler/evidence": {
      "get": {
        "summary": "Retrieve required evidence items and verification statuses",
        "operationId": "getEvidenceList",
        "parameters": [
          {
            "name": "changeSetId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of evidence items",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Finding": {
        "type": "object",
        "required": ["id", "type", "title", "description", "severity", "status"],
        "properties": {
          "id": { "type": "string", "example": "F-001" },
          "type": { "type": "string", "enum": ["BLOCK", "WARNING"] },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "severity": { "type": "string", "enum": ["HIGH", "MEDIUM", "LOW"] },
          "status": { "type": "string", "enum": ["OPEN", "RESOLVED"] },
          "rule": { "type": "string" },
          "affectedEntity": { "type": "string" },
          "obligationId": { "type": "string" }
        }
      },
      "Obligation": {
        "type": "object",
        "required": ["id", "obligation", "owner", "status"],
        "properties": {
          "id": { "type": "string", "example": "OBL-01" },
          "obligation": { "type": "string" },
          "owner": { "type": "string" },
          "status": { "type": "string", "enum": ["OPEN", "COMPLETED", "IN PROGRESS"] },
          "deadline": { "type": "string" },
          "severity": { "type": "string" },
          "findingId": { "type": "string" }
        }
      },
      "EvidenceItem": {
        "type": "object",
        "required": ["id", "title", "status", "buttonText"],
        "properties": {
          "id": { "type": "string", "example": "EVD-01" },
          "title": { "type": "string" },
          "status": { "type": "string", "enum": ["MISSING", "VERIFIED", "AVAILABLE"] },
          "buttonText": { "type": "string" },
          "fileHint": { "type": "string" },
          "fileUrl": { "type": "string", "nullable": true },
          "verifiedBy": { "type": "string", "nullable": true },
          "verificationHash": { "type": "string", "nullable": true }
        }
      },
      "ReadinessSummary": {
        "type": "object",
        "required": ["protocol", "changeSet", "sites", "participants", "blockingFindings", "warnings", "evidence", "status"],
        "properties": {
          "protocol": { "type": "string" },
          "changeSet": { "type": "string" },
          "sites": { "type": "integer" },
          "participants": { "type": "integer" },
          "blockingFindings": { "type": "integer" },
          "warnings": { "type": "integer" },
          "evidence": { "type": "string" },
          "status": { "type": "string", "enum": ["BLOCKED", "READY"] }
        }
      },
      "CompilationStep": {
        "type": "object",
        "required": ["name", "status"],
        "properties": {
          "name": { "type": "string", "enum": ["CHANGESET", "IMPACT", "COMPILE", "RULES", "FINDINGS", "OBLIGATIONS", "EVIDENCE", "VERIFICATION", "READY"] },
          "status": { "type": "string", "enum": ["PENDING", "RUNNING", "PASSED", "FAILED"] }
        }
      },
      "CompilationRunRequest": {
        "type": "object",
        "required": ["changeSetId", "trialId"],
        "properties": {
          "changeSetId": { "type": "string" },
          "trialId": { "type": "string" },
          "triggeredBy": { "type": "string" },
          "pipeline": { "type": "string" }
        }
      },
      "CompilationRunData": {
        "type": "object",
        "required": ["runId", "changeSetId", "trialId", "protocol", "timestamp", "status", "readiness", "pipelineSteps", "auditHash"],
        "properties": {
          "runId": { "type": "string", "example": "CMP-000128" },
          "changeSetId": { "type": "string", "example": "CS-0001" },
          "trialId": { "type": "string", "example": "ATF-001" },
          "protocol": { "type": "string", "example": "v1.1" },
          "timestamp": { "type": "string" },
          "status": { "type": "string", "enum": ["PASSED", "FAILED", "RUNNING", "PENDING"] },
          "readiness": { "$ref": "#/components/schemas/ReadinessSummary" },
          "pipelineSteps": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CompilationStep" }
          },
          "auditHash": { "type": "string" }
        }
      },
...
      "CompilationRunResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/CompilationRunData" },
          "timestamp": { "type": "string" }
        }
      },
      "FindingsResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Finding" } },
          "timestamp": { "type": "string" }
        }
      },
      "ObligationsResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Obligation" } },
          "timestamp": { "type": "string" }
        }
      },
      "EvidenceResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/EvidenceItem" } },
          "timestamp": { "type": "string" }
        }
      },
...
      "ApiErrorEnvelope": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "required": ["code", "message", "timestamp"],
            "properties": {
              "code": { "type": "string", "example": "COMPILATION_BLOCKED" },
              "message": { "type": "string", "example": "Build Failed: 3 Blockers found" },
              "details": {
                "type": "array",
                "items": { "type": "string" }
              },
              "timestamp": { "type": "string" }
            }
          }
        }
      }
    },
    "responses": {
...
      "UnprocessableEntityError": {
        "description": "Validation or business rule error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" }
          }
        }
      }
    }
  }
}